fiam_2025_sentiment_analysis/requirements_optimized.txt
polars>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.35.0
pyarrow>=14.0.0
pyyaml>=6.0
fastparquet>=2023.10.0

